Remove double from price and weight

+++ tkpdtransaction/src/main/java/com/tokopedia/transaction/checkout/view/adapter/SingleAddressShipmentAdapter.java
@@ -468,7 +468,7 @@ public class SingleAddressShipmentAdapter extends RecyclerView.Adapter<RecyclerV
         }
 
         private String getPriceFormat(double pricePlan) {
-            return pricePlan == 0 ? "-" : "Rp" + String.valueOf(pricePlan);
+            return pricePlan == 0 ? "-" : "Rp" + String.valueOf((int) pricePlan);
         }
 
         private String getTotalItemFormatted(int totalItem) {
@@ -476,7 +476,7 @@ public class SingleAddressShipmentAdapter extends RecyclerView.Adapter<RecyclerV
         }
 
         private String getTotalWeightFormatted(double totalWeight) {
-            return String.format("Ongkos Kirim (%s gr)", totalWeight);
+            return String.format("Ongkos Kirim (%s gr)", (int) totalWeight);
         }
 
         private int getExpandOtherProductVisibility(List<CartItemModel> cartItemModels) {
